Now using single precision (float) instead of double precision in the noise signal level detector. Patch from DH1DM.
git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@1644 f8843c6b-282e-0410-804b-c0d5895947d2

+++ src/svxlink/trx/SigLevDetNoise.h
@@ -142,9 +142,9 @@ class SigLevDetNoise : public SigLevDet
      * @brief 	Read the latest calculated signal level
      * @return	Returns the latest calculated signal level
      */
-    double lastSiglev(void) const
+    float lastSiglev(void) const
     {
-      return offset - slope * log10(last_siglev);
+      return offset - slope * log10f(last_siglev);
     }
      
     void reset(void);
@@ -155,7 +155,7 @@ class SigLevDetNoise : public SigLevDet
   private:
     Async::AudioFilter	  *filter;
     Async::SigCAudioSink  *sigc_sink;
-    double    	      	  last_siglev;
+    float    	      	  last_siglev;
     float     	      	  slope;
     float     	      	  offset;
